backend/requirements.txt
Django==5.2.7
django_countries==7.6.1
django_phonenumber_field==8.3.0
djangorestframework==3.16.1
